Fold remaining Editor unit into EditorController
