+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -16,7 +16,6 @@
 #include "Window.h"
 
 #include <openrct2/Context.h>
-#include <openrct2/Editor.h>
 #include <openrct2/Game.h>
 #include <openrct2/GameState.h>
 #include <openrct2/Input.h>
