+++ src/openrct2-ui/windows/EditorScenarioOptions.cpp
@@ -14,7 +14,6 @@
 
 #include <algorithm>
 #include <openrct2/Context.h>
-#include <openrct2/Editor.h>
 #include <openrct2/Game.h>
 #include <openrct2/GameState.h>
 #include <openrct2/OpenRCT2.h>
