--- src/openrct2/Editor.h
@@ -1,32 +1,0 @@
-/*****************************************************************************
- * Copyright (c) 2014-2026 OpenRCT2 developers
- *
- * For a complete list of all authors, please refer to contributors.md
- * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
- *
- * OpenRCT2 is licensed under the GNU General Public License version 3.
- *****************************************************************************/
-
-#pragma once
-
-#include "core/StringTypes.h"
-#include "localisation/StringIdType.h"
-#include "object/ObjectTypes.h"
-
-struct ResultWithMessage;
-
-namespace OpenRCT2
-{
-    struct WindowBase;
-}
-
-namespace OpenRCT2::Editor
-{
-    void ObjectListLoad();
-    ResultWithMessage CheckPark();
-    std::pair<ObjectType, StringId> CheckObjectSelection();
-
-    uint8_t GetSelectedObjectFlags(ObjectType objectType, size_t index);
-    void ClearSelectedObject(ObjectType objectType, size_t index, uint32_t flags);
-    void SetSelectedObject(ObjectType objectType, size_t index, uint32_t flags);
-} // namespace OpenRCT2::Editor
