+++ src/openrct2/rct1/S4Importer.cpp
@@ -10,7 +10,6 @@
 #include "../Cheats.h"
 #include "../Context.h"
 #include "../Diagnostic.h"
-#include "../Editor.h"
 #include "../Game.h"
 #include "../GameState.h"
 #include "../ParkImporter.h"
